+++ src/Preprocess/Preprocess.jl
@@ -1,0 +1,20 @@
+module Preprocess
+
+using StaticArrays
+
+using LinearAlgebra
+using SparseMatricesCSR
+# using ThreadedSparseCSR
+
+using Adapt
+using XCALibre.Multithread
+using XCALibre.Mesh
+using XCALibre.Fields
+using XCALibre.Calculate
+using XCALibre.Discretise
+using XCALibre.ModelPhysics
+# using XCALibre.Turbulence
+
+include("setFields.jl")
+
+end