+++ src/XCALibre.jl
@@ -22,6 +22,7 @@ include("ModelPhysics/ModelPhysics.jl")
 include("Simulate/Simulate.jl")
 include("Solvers/Solvers.jl")
 include("Postprocess/Postprocess.jl")
+include("Preprocess/Preprocess.jl")
 include("Mesh/BlockMesher2D/BlockMesher2D.jl")
 
 using Reexport
@@ -37,6 +38,7 @@ using Reexport
 @reexport using XCALibre.Simulate
 @reexport using XCALibre.Solvers
 @reexport using XCALibre.Postprocess
+@reexport using XCALibre.Preprocess
 @reexport using XCALibre.IOFormats
 @reexport using XCALibre.UNV3
 @reexport using XCALibre.UNV2